+++ pykafe/analytic/templates/wagalytics/dashboard.html
