+++ pykafe/analytic/wagtail_hooks.py
@@ -1,0 +1,59 @@
+
+from wagtail.contrib.modeladmin.options import (
+    ModelAdmin,
+    ModelAdminGroup,
+    modeladmin_register
+)
+from tracking.models import Visitor
+from django.utils.translation import ugettext_lazy as _
+
+try:
+    from wagtail.core import hooks
+    from wagtail.admin.menu import MenuItem
+except ImportError:  # fallback for Wagtail <2.0
+    from wagtail.wagtailcore import hooks
+    from wagtail.wagtailadmin.menu import MenuItem
+
+try:
+    from django.urls import reverse
+except ImportError:  # fallback for Django <1.9
+    from django.core.urlresolvers import reverse
+
+try:
+    from django.urls import re_path, include
+except ImportError:  # fallback for Django <2.0
+    from django.conf.urls import url as re_path
+    from django.conf.urls import include
+
+from analytic import urls
+
+
+@hooks.register('register_admin_urls')
+def register_admin_urls():
+    return [
+        re_path(r'^analytics/', include(urls)),
+    ]
+
+
+class VisitorAdmin(ModelAdmin):
+    model = Visitor
+    menu_label = "Visitor"
+    menu_icon = "pick"
+    menu_order = 300
+    add_to_settings_menu = False
+    exclude_from_explorer = False
+    list_display = ('session_key', 'user', 'start_time', 'ip_address', 'user_agent')
+    list_filter = ('user', 'ip_address')
+    search_fields = ('user',)
+
+
+@hooks.register('register_admin_menu_item')
+def register_styleguide_menu_item():
+    return MenuItem(
+        _('Analytics'),
+        reverse('analytic'),
+        classnames='icon icon-fa-bar-chart',
+        order=1000
+    )
+
+modeladmin_register(VisitorAdmin)
